cloud run code/requirements.txt
# https://pypi.org/project/flask
Flask

#https://pypi.org/project/python-binance/
python-binance

# https://pypi.org/project/gunicorn
gunicorn